include: mm: Update read-ahead values
include: mm: Update read-ahead values

Technically there is a maximum of 256 pages available to the system and each page takes up ~4kb. So _actual_ max page value is 1024kb. Can actually go higher but there is a significant lag in the system when a mis-cache happens with more than 1024kb read ahead.